# Conflicts:
#	assets/tiles/MPG_pixel_adventure.tiled-session
#	lib/components/level_components/entity/isometric_player.dart
#	lib/components/level_components/entity/player.dart
#	lib/core/level/game_world.dart
#	lib/core/physics/collision_block.dart
#	lib/mpg_pixel_adventure.dart

+++ lib/core/physics/collision_block.dart
@@ -9,8 +9,8 @@ import 'package:mpg_achievements_app/mpg_pixel_adventure.dart';
 class CollisionBlock extends PositionComponent
     with CollisionCallbacks, HasGameReference<PixelAdventure> {
   //position and size is given and passed in to the PositionComponent with super
-  int? zPosition;
-  int? zHeight;
+  double? zPosition;
+  double? zHeight;
 
   bool? isIsometric;
   CollisionBlock({
